Move extension method into NpgsqlSnakeCaseNameTranslator
